pyproject.toml
[project]
name = "klanker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiostream>=0.7.1",
    "fastmcp>=3.1.1",
    "jsonschema>=4.26.0",
    "openai>=2.32.0",
    "openrouter>=0.7.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-frontmatter>=1.1.0",
]
